Fix Hibernate Validator metadata split
