+++ tests/src/org.hibernate.validator/hibernate-validator/9.1.0.Final/src/test/java/hibernate_validator/Dto.java
@@ -4,7 +4,7 @@
  * You should have received a copy of the CC0 legalcode along with this
  * work. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.
  */
-package org.hibernate.validator;
+package hibernate_validator;
 
 import jakarta.validation.constraints.AssertFalse;
 import jakarta.validation.constraints.AssertTrue;
